yolo_project/requirements.txt
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
PyYAML>=6.0
streamlit>=1.30.0
streamlit-webrtc>=0.47.0
onnx>=1.15.0
onnxruntime>=1.16.0
tqdm>=4.66.0
pytest>=7.4.0
